NONE: refactor settings, create SettingsPageContent component, move help functions to utils
