+++ renderer/src/components/SettingsPageContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SettingsPageContent";
